requirements.txt
scikit-learn
Pillow
scipy
progressbar
numpy
torch
torchvision
progressbar2
xgboost